server/requirements-rocm.txt
# # apt-get update
# # apt-get install -y emacs mlocate wget git curl gcc g++ make unzip
# # wget https://repo.anaconda.com/archive/Anaconda3-2022.10-Linux-x86_64.sh
# # bash Anaconda3-2022.10-Linux-x86_64.sh

--extra-index-url https://download.pytorch.org/whl/rocm5.4.2
torch==2.0.1
torchaudio==2.0.2
faiss-cpu

tensorboard
tensorboardx

onnxscript
onnxsim
onnxruntime-gpu
torchcrepe
local_attention
pyworld
transformers
safetensors
